refactor(container): restructure component files
