+++ src/shared/ui/container/index.ts
@@ -1,1 +1,1 @@
-export { Container } from './container'
+export { Container } from './ui/container'
